Extract PatientIdentifierSelector as reusable component with localStorage caching
- Create new PatientIdentifierSelector component consolidating duplicate patient search code
- Add localStorage caching for search type preference (key: patient_identifier_filter_search_type)
- Refactor CreateDispenseSheet to use PatientIdentifierSelector (~607 → ~188 lines)
- Refactor CreateMedicationReturnSheet to use PatientIdentifierSelector (~642 → ~244 lines)
- Support callbacks: onPatientSelect, onClearPatient, onRegisterNewPatient (optional)
- Support preselectedPatientId prop for pre-populating patient selection
- Include PatientIDScanDialog and verification dialog in component

Closes #15658
